refactor(mocks): restructure mocking setup and enhance API handlers
