+++ src/components/providers.tsx
@@ -2,7 +2,6 @@
 
 // Since QueryClientProvider relies on useContext under the hood, we have to put 'use client' on top
 import { isServer, QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import { MockServiceWorkerProvider } from './msw-provider'
 
 function makeQueryClient() {
   return new QueryClient({
@@ -40,9 +39,5 @@ export default function Providers({ children }: { children: React.ReactNode }) {
   //       render if it suspends and there is no boundary
   const queryClient = getQueryClient()
 
-  return (
-    <MockServiceWorkerProvider>
-      <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
-    </MockServiceWorkerProvider>
-  )
+  return <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
 }
