+++ src/main.tsx
@@ -1,14 +1,23 @@
 import { router } from '@/router'
 import { RouterProvider } from '@tanstack/react-router'
 import React from 'react'
-import ReactDOM from 'react-dom/client'
+import { createRoot } from 'react-dom/client'
 import Providers from './components/providers'
 import './index.css'
+import { enableMocking } from './mocks/setup'
 
-ReactDOM.createRoot(document.getElementById('root')!).render(
-  <React.StrictMode>
-    <Providers>
-      <RouterProvider router={router} />
-    </Providers>
-  </React.StrictMode>,
-)
+function renderApp() {
+  createRoot(document.getElementById('root')!).render(
+    <React.StrictMode>
+      <Providers>
+        <RouterProvider router={router} />
+      </Providers>
+    </React.StrictMode>,
+  )
+}
+
+enableMocking()
+  .catch((error) => {
+    console.error('Mocking failed', error)
+  })
+  .finally(renderApp)
